feat: move relation validation tests to dedicated file
