+++ tests/unit/validation/validation.test.js
@@ -5,10 +5,6 @@ import fs from "fs";
 import path from "path";
 import {  
     entitiesWithoutPK,
-    relationsUnconnected,
-    validateGraph, 
-    cardinalitiesNotValid,
-    notNMRelationsWithAttributes,
     weakEntitiesWithPrimaryKey,
     weakEntitiesWithoutPartialKey,
     weakEntitiesWithMoreThanOnePartialKey,
@@ -21,88 +17,6 @@ import {
 } from "../../../src/utils/validation"
 
 let graph;
-
-describe("Relations", () => {
-    test("Every relation should connect two entities (can be the same at both sides)", () => {
-        // Ensure the graph is valid initially
-        expect(relationsUnconnected(graph)).toBe(false);
-
-        const initializedSide = { 
-            cardinality: "",
-            cell: "",
-            entity: {
-                idMx: "",
-            },
-            idMx: "",
-        }
-        // Remove attributes from an entity
-        graph.relations.at(1).side1 = initializedSide;
-        graph.relations.at(1).side2 = initializedSide;
-        expect(relationsUnconnected(graph)).toBe(true);
-        expect(validateGraph(graph).noUnconnectedRelations).toBe(false)
-    });
-
-    test("Cant be relations with attributes if they are not N:M", () => {
-        // Ensure the graph is valid initially
-        expect(relationsUnconnected(graph)).toBe(false);
-
-        const attributes = [
-            {
-                "idMx":"9",
-                "name":"Atributo",
-                "position":{
-                    "x":560,
-                    "y":130
-                },
-                "cell":[
-                    "9",
-                    "10"
-                ]
-            },
-        ]
-        // Remove attributes from an entity
-        graph.relations.at(1).attributes = attributes
-        expect(notNMRelationsWithAttributes(graph)).toBe(true);
-    });
-
-    test("Every relation should have valid cardinalities", () => {
-        // Ensure the graph is valid initially
-        expect(cardinalitiesNotValid(graph)).toBe(false);
-
-        const initializedSide1 = { 
-            cardinality: "",
-            cell: "20",
-            entity: {
-                idMx: "",
-            },
-            idMx: "",
-        }
-        const initializedSide2 = { 
-            cardinality: "",
-            cell: "24",
-            entity: {
-                idMx: "",
-            },
-            idMx: "",
-        }
-        // Remove attributes from an entity
-        graph.relations.at(1).side1 = initializedSide1;
-        graph.relations.at(1).side2 = initializedSide2;
-        expect(cardinalitiesNotValid(graph)).toBe(true);
-        expect(validateGraph(graph).noNotValidCardinalities).toBe(false)
-    });
-
-    test("A fully mandatory 1:1 relation should be valid", () => {
-        graph.relations.at(1).side1.cardinality = "1:1";
-        graph.relations.at(1).side2.cardinality = "1:1";
-
-        const diagnostics = validateGraph(graph);
-
-        expect(cardinalitiesNotValid(graph)).toBe(false);
-        expect(diagnostics.noNotValidCardinalities).toBe(true);
-        expect(diagnostics.isValid).toBe(true);
-    });
-});
 
 describe("Weak entities", () => {
     test("an identifying relation is invalid if the strong side is not 1:1", () => {
